add references/Coquelicot_a_user-friendly_library_of_real_analysis_for_coq.pdf
Sylvie Boldo, Catherine Lelay, Guillaume Melquiond. Coquelicot: A
User-Friendly Library of Real Analysis for Coq. 2013. <hal-00860648v1>

Abstract : Real analysis is pervasive to many applications, if only
because it is a suitable tool for modeling physical or socio-economical
systems. As such, its support is warranted in proof assistants, so that
the users have a way to formally verify mathematical theorems and
correctness of critical systems. The Coq system comes with an
axiomatization of standard real numbers and a library of theorems on
real analysis. Unfortunately, this standard library is lacking some
widely used results. For instance, power series are not developed
further than their definition. Moreover, the definitions of integrals
and derivatives are based on dependent types, which make them especially
cumbersome to use in practice. To palliate these inadequacies, we have
designed a user-friendly library: Coquelicot. An easier way of writing
formulas and theorem statements is achieved by relying on total
functions in place of dependent types for limits, derivatives,
integrals, power series, and so on. To help with the proof process, the
library comes with a comprehensive set of theorems that cover not only
these notions, but also some extensions such as parametric integrals,
two-dimensional differentiability, asymptotic behaviors. It also offers
some automations for performing differentiability proofs. Moreover,
Coquelicot is a conservative extension of Coq's standard library and we
provide correspondence theorems between the two libraries. We have
exercised the library on several use cases: in an exam at university
entry level, for the definitions and properties of Bessel functions, and
for the solution of the one-dimensional wave equation.